refactor: (WIP) split encrypt_decrypt into types
