+++ securedrop-protocol/bench/benches/manual.rs
@@ -5,10 +5,11 @@ use rand_chacha::ChaCha20Rng;
 use rand_core::{RngCore, SeedableRng};
 use serde::Serialize;
 
-use securedrop_protocol_minimal::encrypt_decrypt::{
+use securedrop_protocol_minimal::types::{
     Envelope, FetchResponse, Journalist, Plaintext, ServerMessageStore, Source, User,
-    compute_fetch_challenges,
 };
+
+use securedrop_protocol_minimal::encrypt_decrypt::compute_fetch_challenges;
 
 use securedrop_protocol_bench::{bench_decrypt, bench_encrypt, bench_fetch};
